New movement
User clicks on tiles to build a path starting from the 4 tiles surrounding the player's current position.

Tiles in red are pending. Right click the last tile you selected to de-select it. Press the button on the left side to confirm moves.

Next is to implement "turns" so that we can add multiple characters.